refactor shell utils and add skeleton for linter
